Enqueue admin.js regardless of nlu service status.
