+++ webpack.config.js
@@ -8,11 +8,12 @@ module.exports = {
 		path: path.resolve(__dirname, 'dist'),
 	},
 	entry: {
-		editor: [ './src/js/editor.js'],
-		'editor-ocr': [ './src/js/editor-ocr.js'],
-		media: [ './src/js/media.js'],
-		admin: [ './src/js/admin.js'],
-		'gutenberg-plugin': [ './src/js/gutenberg-plugin.js'],
+		editor: [ './src/js/editor.js' ],
+		'editor-ocr': [ './src/js/editor-ocr.js' ],
+		media: [ './src/js/media.js' ],
+		admin: [ './src/js/admin.js' ],
+		'language-processing`': [ './src/js/language-processing.js' ],
+		'gutenberg-plugin': [ './src/js/gutenberg-plugin.js' ],
 		'recommended-content-block': [ './includes/Classifai/Blocks/recommended-content-block/index.js' ],
 		'recommended-content-block-frontend': [ './includes/Classifai/Blocks/recommended-content-block/frontend.js' ],
 	},
